refactor(git): align React module boundaries
